NCL-1294: fix for literal values getting clobbered
In the case where a literal value is only initialized as a parameter to a func or proc within a func or proc, 
the value was getting clobbered prematurely. I don't fully understand the RemapParameters code, but I took the most conservative option I could find 
to fix this. It seems to work now, even with more quirky situations than the one the user came up with.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12800 10fe2890-8a40-4696-8b1d-7ccbb19c398e