NCL-1190: fix for seg fault writing null-valued attribute.
Modified NclHDFEOS.c to include 19 new projection definitions. Additionally if an unknown projection integer id is encountered, the code sets the Grid variable 
projection attribute to "unknown". This fixes the immediate problem. Really, although empty strings can be valid attribute values,  NULL values should not 
be possible in general. Since the projection attribute is a synthetic added attribute, the fault was in the NclHDFEOS code. Even so, I added a bit of
code to the _NclStringMDToCharMD code that issues a warning if a NULL value is encountered and replaces it with a missing value.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12802 10fe2890-8a40-4696-8b1d-7ccbb19c398e