NCL-1299: updating GRIB2 code tables
This adds 12 new tables with 74 parameters. It also adds 
135 new parameters to existing tables. Note that some of these do not
have short names yet (they have the text "Validation" in the space for the short name --presumably meaning that something could change about these variables). 
They will show up as VAR_[some combination of numbers and underscores] in place of the usual
short name prefix.
Remember we are no longer in the business of creating short names.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13177 10fe2890-8a40-4696-8b1d-7ccbb19c398e