NCL-1355: Fix for writing NetCDF4 variables with coordinate variables that contain _FillValue attributes
NetCDF4 does not allow the _FillValue attribute to be added or modified after the variable data has been written.
NCL's traditional code for the "simple" method of writing variables had always added attributes after writing 
the variable. In 2009 I fixed this for the variables themselves, but the NCL code also implicitly creates 
coordiate variables. Since coordinate variables usually need to be completely defined, they do not usually need a
_FillValue attribute. Thus the fact that these were still being created in the old way had not shown up as a problem
until recently. This check-in fixes coordinate variable creation to work in the same order as variable creation.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13240 10fe2890-8a40-4696-8b1d-7ccbb19c398e