Added a built-in "rainbow2" color map. This needs to be renamed.
This addresses NCL-1397.

I originally just used the "rainbow" color map that already exists
("rainbow.gp"), but this caused problems for 
some of the C/Fortran examples:

  fatal:CvtStringToCmap:Unable to convert string "rainbow" to ColorMap
  warning:["Resources.c":770]:Unable to set wkColorMap to default value - Using NULL

I copied this colormap exactly, but turned it into a built-in function.

Dave may have a better suggestion for how to handle this. Also, I don't
like the name "rainbow2". I thought about "rainbow_default", but this
seems like a long name.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@13366 10fe2890-8a40-4696-8b1d-7ccbb19c398e