Updated this function to allow leftmost dimensions.
This addresses NCL-1419, and is needed for the ESMF_regrid_w_weights
routine.

It allows you to reshape an array, and also "conform" the
dimensions at the same time.

That is, if x is 100 x 3 x 20 and you do:

   xr = reshape(x,(/4,5/))

then "xr" will be 100 x 3 x 4 x 5. It recognizes that the product "4 x 5" 
is the same as the rightmost dimension of x (20), and so it's the same as
doing:

   xr = reshape(x,(/100,3,4,5/))



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13396 10fe2890-8a40-4696-8b1d-7ccbb19c398e