NCL-1495: fix for problem with vcZeroFLabelOn
The field was not actually a zero field in the sense of all values having zero magnitude.
This meant that VectorPlot still calls the LLU VVECTR routine. It was printing its own 
zero field label because no vectors actually get drawn. The HLU code was not turning off the
zero field unconditionally as it should have (because it draws a text annotation for the
zero field). Note that the HLU code behaves differently from the LLU code when there are
non-zero vectors but they all fall below the minimum magnitude set by the user. The HLU
code never outputs 'zero field' under these conditions, whereas the LLU code does (if zero 
field text is not turned off).
Also fixed a problem with StreamlinePlot thinking a vector field was a zero field when
actually it just was a uniform (non-zero) field.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@13543 10fe2890-8a40-4696-8b1d-7ccbb19c398e