NCL-1498: A couple of fixes for GRIB2 files:
If a statistically processed variable had too many time periods (necessitating division into multiple variables)
it was resulting in an overrun of a while loop and a core dump.

The code for the probability dimension in GRIB2 had not been thoroughly tested because appropriate sample files
were not available. The sample file SREF_CONUS_40km_ensprod_20120621_1500.grib2 provides such a test.  



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@13549 10fe2890-8a40-4696-8b1d-7ccbb19c398e