NCL-1501: fixed little 'x' problem
The problem was that sometime with a color change only 1 point was being drawn at the beginning
or end of a curly vector line. The LLU CURVE routine calls POINT when it is only given a single
point. POINT draws an 'x' at the point location. The fix was to ensure that 2 point at least are
always drawn.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@13579 10fe2890-8a40-4696-8b1d-7ccbb19c398e