NCL-1514: applied patch from Orion Poplawski to get xwd driver to work in 64-bit mode.
I did some basic tests to see that the correct plot is produced on my Mac and on mirage.
Also I verified that the plot was incorrect or invisible prior to the patch. Note, however,
that I have no 32-bit machine to test that it will still work in 32-bit environments.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@13589 10fe2890-8a40-4696-8b1d-7ccbb19c398e