NCL-1409: add resources to control county level line colors
Added resources:
mpCountyLineColor
mpCountyLineDashPattern
mpCountyLineDashSegLenF
mpCountyLineThicknessF

These are ignored if the Medium Res database is not in use. They work for US climate divisions as well.

Also added resources:

mpProvincialLineColor
mpProvincialLineDashPattern
mpProvincialLineDashSegLenF
mpProvincialLineThicknessF

These are simply aliases for the equivalent USState resources to be more 
descriptive for first order international country divisions (e.g. states of Mexico or provinces of Canada). 
If both the USState and the equivalent Provincial
resource are set in the same call, the last one in the list prevails.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13676 10fe2890-8a40-4696-8b1d-7ccbb19c398e