NCL-1526: fix for bug writing coordinate variable to NetCDF
Now when an attribute is written to a variable the dimensionality of the variable is checked: if it is 1D
and it has a coordinate variable whose name is the same as the variable itself, the attribute is also
written to the coordinate variable. This automatically keeps the coordinate variable attributes in sync 
with the variable itself, eliminating the problem described in the NCL-1526 JIRA ticket.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13685 10fe2890-8a40-4696-8b1d-7ccbb19c398e