NCL-1533: problem with map fill
To my knowledge this is the only example that uses the resource "vpUseSegments". 
Segments are a GKS concept that allows a piece of a graphic to be created once, then
saved in a file, and used again in order to save drawing time. Although most of the
HLU view objects support segment drawing, it has hardly ever been used except in the
old interactive tool. 
The problem is that currently the segment facility does not work with high-numbered, (absolute)
color indexes. Since it is so little used there is probably not a lot of point in fixing this --
at least before the 6.1.0 release..
I am "fixing" these scripts by commenting out the line that sets the "vpUseSegments" resource.

Note that the map colors in the original script were "accidental": the color indexes are the
defaults used by the MapPlot and intended to be used with the default color table, but in this
script they point to colors of the color table used by the plot objects. I have put those old
color indexes into the res file, but commented out. Uncomment it to see the plot as it used to
look.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13717 10fe2890-8a40-4696-8b1d-7ccbb19c398e