NCL-1488: list-style delete now works with attributes
This changes how attributes added to lists behave. Their values now point to the same space as the original. 
Among the implications are that if you change the value in the list it will also change the original 
attribute value. Or if you insert the same attribute into multiple lists, anything you do to the attribute in
one list also happens in the other list (as well as to the original attribute).Along the way fixed a bug in the callback that adjusts the list items when one of its underlying values is deleted. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13732 10fe2890-8a40-4696-8b1d-7ccbb19c398e