NCL-1444: added an 'opt' logical argument that accepts an attribute "cyclic"
This is to make the wgt_area_smooth function more general. It can now be used for arbitrary data arrays of 
2 or more dimensions. The 2D smoothing occurs along the 2 rightmost dimensions. "cyclic" defaults to True
When true the last element and the first element of the rightmost dimension are considered to be next to
each other. In this case the first and last elements of the Y dimension (2nd from the right), are set to
_FillValue. If "cyclic" is False, the this is also True for the X (rightmost) dimension).

 


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13736 10fe2890-8a40-4696-8b1d-7ccbb19c398e