CL-738: add a new attribute-based create function
Tentatively called "create_graphic", this function allows you to set resources in a 'res'-type
attribute when creating a graphical object.

ret_graphic_objs = create_graphic(names : string array, classname : string, parent : obj or string ("defaultapp","noparent", or "null"), res : logical)

This is to make it easier to fix issues like NCL-738, where you need to be able to dynamically decides whether to specify a resource when creating an object.

It is similar in design to 'attsetvalues', which, by the way, I think should have a different name (as a alias).

If we also did a similar thing for getvalues, then we could deprecate the whole viz block syntax.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13741 10fe2890-8a40-4696-8b1d-7ccbb19c398e