NCL-1543: fix for fill drop out
The calculation of the grid boundary for the purpose of filling the area outside the grid, which is controlled by the 
cnGridBoundFill.. resources, has to be done more carefully when these resources are enabled. Now whenever the fill color
is set non-transparent and also the fill pattern is not "hollow fill", the LLU parameter PIT is set to its smallest 
recommended value (0.001). This results in a more careful generation of the grid boundary and fixes the fill drop out that 
appeared in this test case. Note that the PIT parameter is normally settable by the user using the HLU resource 
cnMaxPointDistance, but in this situation the value will be forced. I will note this in the documentation.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13750 10fe2890-8a40-4696-8b1d-7ccbb19c398e