ESMF regridding code will now return type float if float is input.
If this is not desired, set ReturnDouble attribute to True.

This is only appropriate for ESMF_regrid and ESMF_regrid_with_weights.

The lat/lon are still returned as doubles.

This addresses NCL-1558, and I'm admittedly very nervous about this
change, because I jumped through some hoops to keep from having
to make a float copy of the array.

I also fixed a bug where the source input dimensions on the description
file are not the same as the original source dimensions. This is a possible
situation, so I removed the code that was checking if the dimensions were the
same.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13845 10fe2890-8a40-4696-8b1d-7ccbb19c398e