NCL-210: NhlGetIsolines code for triangular mesh contour plots
This was a little more difficult than the standard contour code, because in the
case of contours that overlay maps, the coordinates are pre-projected to allow Schewchuk's code
to handle them as 2d rather than in a spherical domain. That means they need to be converted
back to lat/lon before passing them on to the user. This is complicated by the the HLU polyline code
that does not want big jumps over the cyclic point. The code I have implemented handles the 
situation in the cases I have tested, but there may be situations where it does not quite work.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13993 10fe2890-8a40-4696-8b1d-7ccbb19c398e