NCL-210: implements get_isolines which allows a user to retrieve the points that define a contour line This routine has two names: NhlGetIsoLines and get_isolines.
The signature is get_isolines(contour_plot_id: graphic, levels: numeric (or the special string "plot"))
If the second argument is "plot" it means get all levels that would be drawn with the current setting
of cnLevelSelectionMode. If levels is a scalar number, the isoline for that particular level is returns
as a 2D array variable with a printVarSummary that looks like this:

Variable: L0
Type: float
Total Size: 896 bytes
            224 values
Number of Dimensions: 2
Dimensions and sizes:   [yx | 2] x [points | 112]
Coordinates:
Number Of Attributes: 5
  _FillValue :  9.96921e+36
  level :        0
  segment_count :       4
  start_point : ( 0, 28, 56, 84 )
  n_points :    ( 28, 28, 28, 28 )

where the segment_count gives the number of discreet line segments that make up the isoline, start_point is 
an array attribute that points to the beginning point of each segment and n_points gives the number of points
in each segment.

If the levels argument is an array, the return value is a list of array variables that each appear like the variable
above. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@13994 10fe2890-8a40-4696-8b1d-7ccbb19c398e