NCL-1572: fix for lack of independence for some High/Low resources
Originally this was an LLU Conpack/Conpackt issue. There were no parameters for 
controlling the High/Low label background color and the background perimeter color.
Mods to the HLU[CP/CT]CHHL function (use-modifiable func called from the LLU Fortran)
allowed the color to be changed independently except that it did not seem possible to
make the High and Low labels independently transparent (because this was simulated by
turning off the background or perimeter entirely -- meaning that both had to be transparent
or non-transparent). Now however, with the transparent capabilities provided by _NhlSet[Line/Fill]Opacity
it is possible to truly make these resources independently controllable.
Which is what this check-in does.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14071 10fe2890-8a40-4696-8b1d-7ccbb19c398e