NCL-634, NCL-795 Implement non-staggered rotated GRIB1 GDS grid 205
This adds support for GRIB1 GDS grid 205, described as "Arakawa Non-E Staggered rotated Latitude/Longitude Grid".
The description seems a bit ambiguous. In fact it seems to be non-staggered and 'non-E' meaning an A-type grid really.
I have also implemented the equivalent GRIB2 grid, but due to what seems like an error in the only sample file I have,
it does not work properly with the sample file. I worked around it but it means the coordinates are not as accurate as 
in the GRIB1 case. Therefore I am waiting to check in a GRIB2 solution until I get a more recent file that hopefully 
describes the grid properly.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14105 10fe2890-8a40-4696-8b1d-7ccbb19c398e