NCL-795: adding support for the "non-E staggered rotated" lat/lon grid in GRIB2
This is the GRIB2 version of GRIB1 GDS type 205. It uses the grid template 32769.
center lat/lon values are positioned in the center of the grid, I was able to 
infer the upper right coordinates based on the idea that in the rotated coordinate
space the upper right coordinates are just the negative of lower left coordinates.
The GRIB2 number now match the GRIB1 numbers to 5 significant digits and plots of
the same variable look indistinguishable to me.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14112 10fe2890-8a40-4696-8b1d-7ccbb19c398e