Added JIRA1530 macro to flag the build of a special GKS library.
This library will have the XQuartz/cairo fix compiled for the X11
driver issue mentioned in NCL-1530. 

This special library is built only if the following is included in
the Site.local file:

#define JIRA1530 1

Really, the only thing that's happening is "cro.c" is built
with -D__JIRA1530__ for this special library.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14121 10fe2890-8a40-4696-8b1d-7ccbb19c398e