Added the DEPXXXX macros needed for the special GKS library build.
This partially addresses NCL-1530. If the following line is included
in the Site.local file:

#define JIRA1530 1

then a special GKS library will be built where "cro.c" is compiled
with "-D__JIRA150__", and a special "ncl.xq.fix" binary will be built
against this GKS library. The regular "ncl" binary will still be built.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14123 10fe2890-8a40-4696-8b1d-7ccbb19c398e