NCL-1630: fix to cd_calendar and cd_inv_calendar
This is peripheral to the JIRA ticket, but there was
a global error flag in nctime that, once set to indicate an error,
was never reset to the initial value. This meant that if one call
had an error, say an invalid units string, all subsequent calls
return missing values, regardless of their validity. This is very
inconvenient when running interactively in particular.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14168 10fe2890-8a40-4696-8b1d-7ccbb19c398e