NCL-1687:
This fix is due to Jonathon Vigh's problem.
Where even the user has used $$ sign to refer
to a varible with special character (here "."),
but some how NCL still treat it as refer to
a component of compound.

The current fix is that we check the whole
string first. If we have the whole string
as a variable, then we will read it.
Otherwise, we will check further.

Also, since we are (at least trying to) handle
compound in the advanced file structure.
We should not have any compound data come to
NclFile.c anyway.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14231 10fe2890-8a40-4696-8b1d-7ccbb19c398e