Added 2 utility color functions:
rgba_to_color_index
color_index_to_rgba

color_index_to_rgba allows you to discover the RGBA values when you
do a getvalues on a color index array set with high numbered values.
For now it does not work on low-numbered color indexes - 0 - 255 ; it
would need to have a workstation argument to do that. It should probably not
be advertised at the moment.

rgba_to_color_index allows to get a high numbered integer that you can use
to set a color index type to a particular color without reference to a color table.
This is more convenient than using a 4-tuple of floats when you need to
set a big array with the color indexes of a large number of poly[gons/lines] as
you need to do with the new poly resources.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14244 10fe2890-8a40-4696-8b1d-7ccbb19c398e