NCL-1695: fix for seemingly odd separation of variables into 2 parts introduced in 6.1.0-beta.
The explanation is that in the particular files where this behavior manifested, the GRIB 1 octet
called the "Time Range Indicator" was set to 0 instead of 1 for a number of records. Most of the dataset
has the value 1. (0 indicates forecast product or unitialized analysis product, while 1 indicates an initialized
analysis product)
Formerly these two values had been considered as equal for the purposes of determining whether records could be
part of the same variable. When I implemented changes removing the '_10' suffix (indicating time range indicator 10) I
inadvertenty reversed the sense of the test for equality of this octet. This resulted in the separation of the variable
into 2 parts. I have fixed this test.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14245 10fe2890-8a40-4696-8b1d-7ccbb19c398e