NCL-1696: fixed problem setting mpUSState line resources.
The attempt to add new mpProvincial resources intended as more proper sounding aliases for
the USState line resources to accommodate non-US boundaries within countries was too simplistic.
Now the resources work as intended -- it should be noted that now if both resources are set the 
Provincial version overrides the USStates version. Formerly, the winning resources would just be
the one that was set last. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14251 10fe2890-8a40-4696-8b1d-7ccbb19c398e