Added recognition of two "nearest neighbor" interpolation methods
This partially addresses NCL-1581. You must have ESMF_RegridWeightGen
6.1.0 or later. (5.2.0rp2 is distributed with NCL V6.1.2).

From the ESMF_RegridWeightGen documentation
http://www.earthsystemmodeling.org/esmf_releases/non_public/ESMF_6_1_0/ESMF_refdoc/node3.html#SECTION03020000000000000000

The nearest neighbor interpolation options work by associating a point
in one set with the closest point in another set. If two points are
equally close then the point with the smallest index is arbitrarily
used (i.e. the point with that would have the smallest index in the
weight matrix). There are two versions of this type of interpolation
available in the regrid weight generation application. One of these is
the nearest source to destination method ("-m neareststod"). In this
method each destination point is mapped to the closest source
point. The other of these is the nearest destination to source method
("-m nearestdtos"). In this method each source point is mapped to the
closest destination point. Note, that with this method the unmapped
destination point detection doesn't work, so no error will be returned
even if there destination points which don't map to any source point.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14256 10fe2890-8a40-4696-8b1d-7ccbb19c398e