NCL-1703 - fix for unbounded growth of object table
There were 2 places objects were being leaks. CallJMP_SCALAR_TRUE_OP/CallJMP_SCALAR_FALSE_OP and in NclVar.c:VarWrite. In the JMP cases, an extra object is created when the fast
(short-circuit) branch is taken (an appropriate logical scalar on the left side of .and. and .or.). In VarWrite the _FillValue attribute value was being created with PERMANENT status.
I am not totally clear on this, but changing it to TEMPORARY seemed to have no bad effect when I ran the full test suite, and stopped the attribute value from hanging around after the local variable was deleted. This makes each function call in the inner loop of the test script take a consistent amount of time and allows the script to finish without problems.
I also took the opportunity to increase the size of the NCL object hash table from 8192 to 32767. Fewer collisions result in faster access, resulting in another speedup of 5-10 per cent.

Note that just starting NCL on the command line before loading any optional code causes more than 17000 objects to be created.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14260 10fe2890-8a40-4696-8b1d-7ccbb19c398e