Fixed error pointed out by Andrew Dawson
On 5/29/13 1:20 PM, Andrew Dawson wrote:

I think there is a small error in the source for the subroutine poisxy2,
which is called by poisson_grid_fill. The error is only present for cyclic
grids.

Line 118 of fill_msg_grid.f reads:

    if (i.eq.1  .and. gtype.eq.1) im1 = il-1

I think this should be:

    if (i.eq.1  .and. gtype.eq.1) im1 = il

When determining the longitude points contributing to the Laplacian of a
point on the left-most edge of a cyclic grid, the correct points are the
right-most point at the same latitude and the point to the right of the
current one. The code as it stands uses the second right-most value instead
of the right-most value, probably just a typing error. I hope that makes
sense!



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14292 10fe2890-8a40-4696-8b1d-7ccbb19c398e