NCL-298: This is code that speeds up polygon and polyline drawing.
The biggest speed-up comes from two new resources that allow for
many individually-colored polygons or polylines to be drawn as a single object.
These are both GraphicStyle resources:
gsSegments which contains the index into the the x/y point arrays of the start of each
new poly segment.
gsColors, which applies to polylines and polygons, and which can contain a color index
for each poly segment. In this case it will have the same size as gsSegments. If it contains
fewer elements the values will be recycled as often as required to draw all segments.

The other significant speed-up is primarily focused on polygons drawn over a map. The speed-up
is enabled automatically for qualifying polygons. However, because this fast method is
not suitable for all polygons there is a resource that allows user control called
mpPolyMode. This is a enumerated resource with 3 values:
"AutoPoly", "FastPoly", and "StandardPoly".
The default is "AutoPoly", meaning that the drawing code decides which type of polygon to draw
based (currently) on the great circle distance of each component line segment, and on the polygon's graphic style 
attributes. If the great circle distance of any component line segment is greater than 10 degress or
if any fill pattern other than solid is in effect, the "standard" method that employs the LLU Areas package
is used. Otherwise a faster method that avoids the Areas package is used.
A small (maybe up to 5%) additional speed-up may be obtained for large polygon arrays if "FastPoly" is specified
primarily because the great circle distance check is eliminated.
On the other hand, the user can guarantee that the old  code is used by setting "StandardPoly". This could
be needed if the auto test does not work properly in some situation, as yet undiscovered. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14295 10fe2890-8a40-4696-8b1d-7ccbb19c398e