Updated GetFillColor to allow RGB[A] array or color map name.
This required moving the "read_colormap_file" function to
"utilities.ncl" so that "GetFillColor", which is in "contributed.ncl",
could reference it.

contributed.ncl is a stand-alone script that should not be required
depend on anything in gsn_code.ncl or gsn_csm.ncl, but since
utilities.ncl is loaded automatically, it's okay to use it.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14303 10fe2890-8a40-4696-8b1d-7ccbb19c398e