NCL-1738: adds a new LabelBar resource called lbBoxSeparatorLinesOn
This allows a user to draw a LabelBar with no interior box lines (box separator lines)
but just a perimeter line around the "bar" of the LabelBar. I am a bit conflicted about
this because I am not sure it is the clearest way to distinguish the parts. But it
is still the dev version...




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14344 10fe2890-8a40-4696-8b1d-7ccbb19c398e