Added a "TriangularMesh" option allowing the user to input their own triangular mesh.
This is to address NCL-1778. Keith Oleson had a regridding script that
was calling unstructured_to_ESMF multiple times, and hence calling "csstri"
multiple times. This caused the script to seemingly hang after a certain
number of times.

I think there's an issue with csstri, so for now, I'm allowing the
user to input the triangular mesh, and that way it's only called once.
 


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14499 10fe2890-8a40-4696-8b1d-7ccbb19c398e