Added wrapper for lspoly_n, and a condensed wrf_plotfmt_read function.
The lspolyW.c check-in addresses NCL-658.

I noticed that the "lspoly" documentation mentioned "x" could be 1D,
but the wrapper was implemented to allow any nD array. I decided to
update it to allow multiple situations:

   x (1d), y (nd), wgt (scalar) (rightmost x and y same length)
   x (1d), y (nd), wgt (1d) (x, wgt, rightmost y same length)
   x (nd), y (nd), wgt (scalar) (x and y the same dimensionality)
   x (nd), y (nd), wgt (nd) (all the same dimensionality)

The wrfW.c check-in addresses NCL-1682. It takes the
three routines wrf_plotfmt_open, wrf_plotfmt_rdhead,
and wrf_plotfmt_rddata and condenses them into one
routine, wrf_plotfmt_read.

I'm not sure I like this, because it means now you have a 
3D arrays whose rightmost two dimensions can be different
dimensionalities. Also, the 3D array has several attributes
that are arrays of strings, and one that is a 2D array of floats.

I believe this routine is not going to be exposed to users, so I'm
hoping that by implementing it this way it makes it more accessible
for the wrf_xxx developers.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14580 10fe2890-8a40-4696-8b1d-7ccbb19c398e