Created RGB versions of all our colormaps, including built-in ones.
This partially addresses NCL-1579.

Note that the RGB file versions of the built-in colormaps are not yet
being installed. This is because they are still part of Palette.c, and
because the built-in colormaps have black background, white
foregrounds, whereas if you use an RGB file version, this gets
reversed.

I have to think about this, because I thought we were forcing all
colormap background/foreground to be white/black as of 6.1.x.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14600 10fe2890-8a40-4696-8b1d-7ccbb19c398e