Modification to previous changes made for ncl-1593 and ncl-1667 -- two related issues that described and identified instances where "thin-white" lines sometimes occur in plots as a graphical artifact.
The strategy in the previous fixes was that for filled areas, we not only perform a cairo-fill operation, but we also perform a cairo-stroke around the border. This essentially filled in the thin-lines, but unfortunately also caused filled text-fonts to have a "bloated" appearance. The difficulty was that at the cairo driver level, we had no state to indicate we were performing a fill operation on a text font versus any other shape.

The solution employed here is a hack to the GKS subsystem to distinguish between filling fonts versus any other shape; a new fill-state was added -- SOLID_TEXT_FILL.  Fortunately, upstream of that, there appears to be only one place in plotchar that needs to use this new state, in plotchar/plchhq.f, which is the "high-quality" filled fonts routine. So:

- plotchar/plchhq.f: was modified to use the new fill-state

- libncarg_gks/awi/gsfais.f: was modified to accept the new fill-state as a valid value

- libncarg_gks/gks.h and libncarg_gks/gks.h: hold the definition of the new fill-state

- libncarg_gks/cdrivers/cro.c: modified to act accordingly to the new fill state; if we are drawing text, only
  the cairo-fill operation is performed, and not the cairo-stroke operation.

- all the remaining legacy drivers were modified to treat the new fill-state the same as the regular solid-fill state:

libncarg_gks/bwi/g01fa.f:
libncarg_gks/cdrivers/ctxt.c:
libncarg_gks/cdrivers/pdf.c:
libncarg_gks/cdrivers/pixoutput.c:
libncarg_gks/cdrivers/ps.c:
libncarg_gks/cdrivers/xoutput.c:

(It was not necessary to modify the WISS driver, as it merely copies state variables verbatum, and it is up to some other driver to perform interpretation).




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14620 10fe2890-8a40-4696-8b1d-7ccbb19c398e