relhum no longer requires all input arrays to have _FillValue attributes.
This addresses NCL-1834.

Note that there's a potential incompatible change with this fix.
The return missing value is now whatever the default missing
value is for the type being returned, and *not* t@_FillValue,
as it was in the past.

This is to keep the function clean and less confusing.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14627 10fe2890-8a40-4696-8b1d-7ccbb19c398e