Updated the gsn_xxx_shapefile polyline/gon routlines to use the new gsSegments resource.
This is in reference to NCL-298.

Dave added a new gsSegments resource which means you no longer need to
create a separate GraphicStyle object for every new line or polygon.
I saw a speed-up of 80 times when plotting the FRA_adm5 shapefile
using this resource.

Also a minor fix where "vc_monofillarrowcolor" should be
"vc_monofillarrowfillcolor"



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14667 10fe2890-8a40-4696-8b1d-7ccbb19c398e