NCL-533, NCL-1693: Fixed the HDF-EOS2 grid structure, bad lat/lon values issue.
These are two seperator issues:

	NCL-533, could be an problem form the fiel itself.
In the HDF-EOS document, it is said that the corner coordinate sould be
in the form of: DDDMMMSSS.ss (double), but in the fiel itself,
the corner is DDD.dd.
	To fix this problem, we have added code to check the corner
cooordinate, if it is in DDD.dd, then we do the lat/lon calculation ourself.

	NCL-1693 related to dataset where there is no XDim/YDim,
so we'd better use the coordinate variables in the file
(even the lat/lon look awkward, as there are missing values in lat/lon).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@14668 10fe2890-8a40-4696-8b1d-7ccbb19c398e