Added regCoef_n.
This is to help Adam make his new diagnostics scripts faster (hopefully).

This wrapper is different from the others, as it takes two arrays of
dimension indexes: one for x and one for y.

 rc = regCoef_n(x,y,0,0)

This is potentially because x and y may be ordered differently.

I've tested this new function on a small test case, but I'm not 
fully on board with it yet. Adam can help test with his real code.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15090 10fe2890-8a40-4696-8b1d-7ccbb19c398e