NCL-1525: added fatal error message if drawing over a map and coordinates are out of range. Specifically this applies to attempts to use RasterFill or AreaFill where some coordinate values have been set to missing values. Note that there is no resource that can be set to indicate missing values in ContourPlot coordinates (or in ScalarField/VectorField data objects in general).
The standard rasterfill and areafill algorithms do not work properly with missing coordinate values.
CellFill does work. And so does setting trGridType to "triangularmesh". There should probably be a more
straightforward resource to use to get the desired behavior.

But as of now, if you try to use RasterFill or AreaFill over a map without setting TriangularMesh, and there are
coordinates that fall outside +/- 540 for longitude or +/- 90 for latitude, you will get an error like this:

fatal:ContourPlotDraw: coordinates are out of range for drawing over a map: standard AreaFill rendering method wil not work;
 consider setting the resouce trGridType to "TriangularMesh" if coordinates contain missing values



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15091 10fe2890-8a40-4696-8b1d-7ccbb19c398e