NCL-1903, NCL-1907: Fixed the seg. fault problem.
This check-in fixed these two issues, plus fixed an issue
which Dave got from Greg Thompson.

But, Dave and I both check the change and feel
logically the previous code and the new change
should do the same thing.

So, this worries us as it could be a compiler issue.

Will it bite sometime again?


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15109 10fe2890-8a40-4696-8b1d-7ccbb19c398e