NCL-1684: fix for problem distinguishing records in NCEP RTMA GRIB2 files
Since this came up again, I implemented a fix for GRIB records that differ
only in the value of the "generating process type". The names of the variables
only differ by the suffix _n (1 ... n), but in addition the 
generating_process_type attribute is added to allow the user to distinguish
the variables. This attribute only appears in the case of otherwise indistiguishable
variable names.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15137 10fe2890-8a40-4696-8b1d-7ccbb19c398e