NCL-152, NCL-628, NCL-1134, NCL-285
This addresses most (but not all) issues concerned with filling areas with a constant or near constant field.
By default, NCL behaves as always, but if you set the new resource cnConstFEnableFill to True, then constant
fields will be filled, including fields that are not truly constant but where all values fall between 2 contour levels.
It works for CellFill, RasterFill, and in most cases for AreaFill. If the field is truly constant the constant field
label will still be displayed but it can be turned off using the cnConstFLabelOn resource. Eventually perhaps the
label should be turned off by default if the cnConstFEnableFill is set True, and cnConstFEnableFill should default to 
True. Enabling constant fill for AreaFill uses the known hack of temporarily setting a data value to something outside the
levels bracketing the constant or near-constant value, then temporarily setting the mono fill resources on and setting the
fill and pattern resources to the values assigned for the constant value. NCL must be able to determine that the field is 
(near) constant before drawing for this method to work.

The one case where this does not work is Adam's stippling example (NCL-1134). That is because in this case the field as a whole
is definitely not constant: a global dataset is overlaid on a limited area map. The portion of the dataset that is over the map is 
locally constant, but NCL cannot pre-determine that fact because the global max and min are far apart. 

Also fixed a bug in CellFill plotting of triangular mesh data: cells around the edges were not always drawn. Now they are. 





git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15156 10fe2890-8a40-4696-8b1d-7ccbb19c398e