Added a CtoFLibrariesUser, IncSearchUser, and LibSearchUser for the ncargf77, ncargcc, etc scripts.
This is so the development versions of these macros (CtoFLibraries,
IncSearch, LibSearch), don't end up in these low-level scripts.  

That is, we don't want "/usr/local/lib/libgfortran.a" or
"-L/Users/home/haley/external/lib" showing up in "ncargcc", "nhlf77",
etc.

These *User macros should be as "clean" as possible, with no hard-coded links or weird paths.
For example, you want:

#define CtoFLibrariesUser "-lgfortran  -lm"

and not

#define CtoFLibrariesUser "-L/usr/local/lib -lgfortran  -lm"

Use the CtoFLibraries macro (as you may be doing now) to get funky
with the paths for development purposes.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15164 10fe2890-8a40-4696-8b1d-7ccbb19c398e